+++ Project.toml
@@ -12,6 +12,7 @@ Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 FunSQL = "cf6cc811-59f4-4a10-b258-a8547a8f6407"
 HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
+JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
 LibPQ = "194296ae-ab2e-5f79-8cd4-7183a0a5a0d1"
 MLStyle = "d8e11817-5142-5d16-987a-aa16d5891078"
 MySQL = "39abe10b-433b-5dbd-92d4-e302a9df00cd"
@@ -34,6 +35,7 @@ DBInterface = "2.6.1"
 DataFrames = "1.0"
 Dates = "1.11.0"
 HTTP = "1.10.15"
+JSON3 = "1.14.1"
 MLStyle = "0.4.17"
 MySQL = "1.4.7"
 Reexport = "1.2.2"
